move PyScriptTest into support.py and introduce test_00_support.py which will test only the machinery
